requirements.txt
# ============================================================
# Apex Agent — Python Dependencies
# ============================================================
# Install with:  pip install -r requirements.txt
#
# Sections:
#   Core           — always required
#   Web / async    — Flask UI + WebSocket mobile bridge
#   Monitoring     — system and GPU metrics
#   Security       — encryption and forensic tools
#   Vision         — screen control, OCR, image processing
#   Speech         — voice input and TTS output
#   Cloud          — AWS S3 and Dropbox sync
#   Containers     — Docker management
#   Extras         — optional convenience libraries
# ============================================================

# ── Core ──────────────────────────────────────────────────────────────────────
requests>=2.31.0          # HTTP client — LLM API calls, web scanning, research
pyyaml>=6.0               # config.yaml parsing (imported as yaml)

# ── Web / async ───────────────────────────────────────────────────────────────
flask>=3.0.0              # pico_bridge web UI, desktop dashboard REST API
websockets>=12.0          # mobile bridge WebSocket server

# ── System monitoring ─────────────────────────────────────────────────────────
psutil>=5.9.0             # CPU, memory, disk, process monitoring
gputil>=1.4.0             # GPU utilisation (security monitoring dashboard)

# ── Terminal ──────────────────────────────────────────────────────────────────
colorama>=0.4.6           # cross-platform ANSI colours in terminal output

# ── Security / crypto ─────────────────────────────────────────────────────────
cryptography>=42.0.0      # Fernet encryption, PBKDF2 key derivation (cloud sync,
                          # forensic evidence integrity)

# ── Computer vision and screen control ────────────────────────────────────────
pyautogui>=0.9.54         # screen reading, clicking, typing, scrolling
pillow>=10.0.0            # image capture and processing (PIL)
opencv-python>=4.8.0      # computer vision, image analysis (cv2)
numpy>=1.24.0             # array operations required by OpenCV and EasyOCR
easyocr>=1.7.0            # on-device OCR for screen text extraction

# ── Speech ────────────────────────────────────────────────────────────────────
speechrecognition>=3.10.0 # microphone input and speech-to-text
pyttsx3>=2.90             # offline text-to-speech (no API key required)

# ── Cloud sync ────────────────────────────────────────────────────────────────
boto3>=1.34.0             # AWS S3 cloud sync
dropbox>=12.0.0           # Dropbox cloud sync

# ── Container management ──────────────────────────────────────────────────────
docker>=7.0.0             # Docker SDK — service deployment and management

# ── Extras ────────────────────────────────────────────────────────────────────
pyperclip>=1.9.0          # clipboard read/write fallback (pico bridge)
